+++ eslint.config.js
@@ -1,14 +1,15 @@
-import js from "@eslint/js";
+import eslintJs from "@eslint/js";
+import eslintJson from "@eslint/json";
+import eslintPluginComments from "@eslint-community/eslint-plugin-eslint-comments/configs";
+import eslintPluginStylistic from "@stylistic/eslint-plugin";
 import eslintConfigPrettier from "eslint-config-prettier";
 import eslintPluginArrayFunc from "eslint-plugin-array-func";
 import eslintPluginCompat from "eslint-plugin-compat";
 import eslintPluginDepend from "eslint-plugin-depend";
 import eslintPluginImport from "eslint-plugin-import";
-import eslintPluginJsonc from "eslint-plugin-jsonc";
-import eslintPluginMarkdown from "eslint-plugin-markdown";
+import * as eslintPluginMdx from "eslint-plugin-mdx";
 import eslintPluginNoSecrets from "eslint-plugin-no-secrets";
 import eslintPluginNoUnsanitized from "eslint-plugin-no-unsanitized";
-import eslintPluginPrettier from "eslint-plugin-prettier";
 import eslintPluginPrettierRecommended from "eslint-plugin-prettier/recommended";
 import eslintPluginPromise from "eslint-plugin-promise";
 import eslintPluginRegexp from "eslint-plugin-regexp";
@@ -20,77 +21,39 @@ import eslintPluginUnicorn from "eslint-plugin-unicorn";
 import eslintPluginYml from "eslint-plugin-yml";
 import globals from "globals";
 
-const FLAT_RECOMMENDED = "flat/recommended";
+export const FLAT_ALL = "flat/all";
+export const FLAT_BASE = "flat/base";
+export const FLAT_RECOMMENDED = "flat/recommended";
 
-export default [
-  {
-    ignores: [
-      "!.circleci",
-      "**/__pycache__/",
-      "**/*min.css",
-      "**/*min.js",
-      "*~",
-      ".git/",
-      ".mypy_cache/",
-      ".pytest_cache/",
-      ".ruff_cache/",
-      ".venv/",
-      "dist/",
-      "node_modules/",
-      "package-lock.json",
-      "poetry.lock",
-      "test-results/",
-      "typings/",
-    ],
-  },
-  js.configs.recommended,
-  eslintPluginArrayFunc.configs.all,
-  eslintPluginCompat.configs[FLAT_RECOMMENDED],
-  eslintPluginDepend.configs[FLAT_RECOMMENDED],
-  eslintPluginImport.flatConfigs.recommended,
-  ...eslintPluginJsonc.configs["flat/recommended-with-jsonc"],
-  ...eslintPluginMarkdown.configs.recommended,
-  eslintPluginNoUnsanitized.configs.recommended,
-  eslintPluginPrettierRecommended,
-  eslintPluginPromise.configs[FLAT_RECOMMENDED],
-  eslintPluginRegexp.configs[FLAT_RECOMMENDED],
-  eslintPluginSecurity.configs.recommended,
-  eslintPluginSonarjs.configs.recommended,
-  ...eslintPluginToml.configs[FLAT_RECOMMENDED],
-  ...eslintPluginYml.configs["flat/standard"],
-  ...eslintPluginYml.configs["flat/prettier"],
-  eslintConfigPrettier, // Best if last
-  {
+export const CONFIGS = {
+  js: {
+    ...eslintJs.configs.recommended,
+    ...eslintPluginArrayFunc.configs.all,
+    ...eslintPluginComments.recommended,
+    ...eslintPluginCompat.configs[FLAT_RECOMMENDED],
+    ...eslintPluginDepend.configs[FLAT_RECOMMENDED],
+    ...eslintPluginImport.flatConfigs.recommended,
+    ...eslintPluginNoUnsanitized.configs.recommended,
+    ...eslintPluginPromise.configs[FLAT_RECOMMENDED],
+    ...eslintPluginRegexp.configs[FLAT_RECOMMENDED],
+    ...eslintPluginSonarjs.configs.recommended,
+    ...eslintPluginUnicorn.configs.recommended,
+    plugins: {
+      depend: eslintPluginDepend,
+      "no-secrets": eslintPluginNoSecrets,
+      "simple-import-sort": eslintPluginSimpleImportSort,
+      sonarjs: eslintPluginSonarjs,
+      unicorn: eslintPluginUnicorn,
+    },
     languageOptions: {
       // eslint-plugin-import sets this to 2018.
       ecmaVersion: "latest",
-      globals: {
-        ...globals.node,
-        ...globals.browser,
-      },
-    },
-    linterOptions: {
-      reportUnusedDisableDirectives: "warn",
-    },
-    plugins: {
-      arrayFunc: eslintPluginArrayFunc,
-      jsonc: eslintPluginJsonc,
-      markdown: eslintPluginMarkdown,
-      "no-secrets": eslintPluginNoSecrets,
-      prettier: eslintPluginPrettier,
-      promise: eslintPluginPromise,
-      security: eslintPluginSecurity,
-      "simple-import-sort": eslintPluginSimpleImportSort,
-      toml: eslintPluginToml,
-      unicorn: eslintPluginUnicorn,
-      yml: eslintPluginYml,
     },
     rules: {
       "array-func/prefer-array-from": "off", // for modern browsers the spread operator, as preferred by unicorn, works fine.
       "depend/ban-dependencies": [
         "error",
         {
-          // import-x doesn't work with eslint 9 yet
           allowed: ["eslint-plugin-import"],
         },
       ],
@@ -98,7 +61,6 @@ export default [
       "no-console": "warn",
       "no-debugger": "warn",
       "no-secrets/no-secrets": "error",
-      "prettier/prettier": "warn",
       "security/detect-object-injection": "off",
       "simple-import-sort/exports": "warn",
       "simple-import-sort/imports": "warn",
@@ -112,30 +74,94 @@ export default [
       "unicorn/switch-case-braces": ["warn", "avoid"],
     },
   },
+};
+Object.freeze(CONFIGS);
+
+export default [
   {
-    files: ["**/*.md"],
-    processor: "markdown/markdown",
+    ignores: [
+      "!.circleci",
+      "**/__pycache__/",
+      "**/*min.css",
+      "**/*min.js",
+      "*~",
+      ".git/",
+      ".*cache/",
+      ".venv/",
+      "dist/",
+      "node_modules/",
+      "package-lock.json",
+      "uv.lock",
+      "test-results/",
+      "typings/",
+    ],
+  },
+  eslintPluginPrettierRecommended,
+  eslintPluginSecurity.configs.recommended,
+  eslintPluginStylistic.configs["all-flat"],
+  {
+    languageOptions: {
+      globals: {
+        ...globals.node,
+      },
+    },
+    linterOptions: {
+      reportUnusedDisableDirectives: "warn",
+    },
     rules: {
-      "prettier/prettier": ["warn", { parser: "markdown" }],
+      "prettier/prettier": "warn",
     },
   },
   {
-    files: ["**/*.md/*.js"], // Will match js code inside *.md files
+    files: ["**/*.js"],
+    ...CONFIGS.js,
+  },
+  {
+    files: ["**/*.json", "**/*.md/*.json"],
+    plugins: {
+      json: eslintJson,
+    },
+    ...eslintJson.configs.recommended,
+    language: "json/json",
+  },
+  {
+    files: ["package.json"],
+    languageOptions: {
+      parser: "jsonc-eslint-parser",
+    },
+    plugins: { depend: eslintPluginDepend },
+    rules: {
+      "depend/ban-dependencies": "error",
+    },
+  },
+  {
+    files: ["**/*.{md,mdx}"],
+    ...eslintPluginMdx.flat,
+    ...eslintPluginMdx.flatCodeBlocks,
+    processor: eslintPluginMdx.createRemarkProcessor({
+      lintCodeBlocks: true,
+    }),
     rules: {
       "no-undef": "off",
       "no-unused-vars": "off",
+      "prettier/prettier": ["warn", { parser: "markdown" }],
     },
   },
+  ...eslintPluginToml.configs[FLAT_BASE],
   {
-    files: ["**/*.md/*.sh"],
+    files: ["**/*.toml", "**/*.md/*.toml"],
     rules: {
-      "prettier/prettier": ["error", { parser: "sh" }],
-    },
-  },
-  {
-    files: ["**/*.toml"],
-    rules: {
+      ...eslintPluginToml.configs[FLAT_RECOMMENDED].rules,
       "prettier/prettier": ["error", { parser: "toml" }],
+    },
+  },
+  ...eslintPluginYml.configs[FLAT_BASE],
+  {
+    files: ["**/*.yaml", "**/*.yml", "**/*.md/*.yaml"],
+    rules: {
+      ...eslintPluginYml.configs[FLAT_RECOMMENDED].rules,
+      ...eslintPluginYml.configs["flat/prettier"].rules,
+      "prettier/prettier": ["error", { parser: "yaml" }],
     },
   },
   {
@@ -148,4 +174,5 @@ export default [
       "yml/no-empty-mapping-value": "off",
     },
   },
+  eslintConfigPrettier, // Best if last
 ];